Add autogenerated patches and automated BSP versioning (#4)
This PR adds the auto generated patches generated from our
internal tool (qcom-patch-generator). Some patches are applied
via Yocto bbappends and other patches have to be applied via
our patch script imdt-patch-qcs8550-build.sh.

I have also added recipe for lsb_release which adds version information
to the image based on the manifest being used. The recipe parses the manifest
XML and gets the name and version from it. This means we don't
need to manually update the version information when we do a build.
We just need to make sure we are using the latest release manifest.

Note that this PR also includes support for streaming the AR1335's on
CSI4 and CSI6 (camera slots connected to I2C switch). The camxoverrides
text file has been updated to force the multicamera logical XML to the XML
that has all slots connected to an AR1335. This means streaming should work
out the box for our customers.