Add support for SSH login as root user with password (#7)
This commit adds support for SSH login as root user using a password.
At the moment, you can't log in to the root user account without
setting up ssh keys on the device. This is a bit annoying to do for
development and it's a simple config change to enable root login.
Once projects are out of the development phase they can change
the SSH configuration details.