Enable networking via Ethernet and add PCIe0/1 devicetree patches (#8)
This commit adds support for networking via Ethernet using
systemd. Two systemd network files have been provided
which allow us configure the Ethernet interface with a DHCP
client or server running. This means we can connect the ethernet
interface to a router and communicate with the SBC over our local network.
And we can also connect our Ethernet interface directly to a laptop
and it will assign our laptop an IP. Then we can stream video directly
to our laptop via TCP/UDP with GStreamer (or custom application).

Note that we currently have a slight issue with systemd-networkd
where it fails to set up our interfaces on boot because it's missing
a machine-id in /etc/machine-id. This issue occurs specifically
because our build uses an overlayfs over a read-only (RO) rootfs.
And when systemd is creating the machine-id, the rootfs is read-only
so it creates a temporary file at /run/machine-id which it bind mounts to
/etc/machine-id. However, the overlayfs gets applied later in the boot process
and essentially hides that so we don't see /etc/machine-id. Then we get
networkd errors about failing to configure the DHCP client
when we try to enable our Ethernet interface.

The issue is described more clearly in the following links:
https://bootlin.com/blog/systemd-read-only-rootfs-and-overlay-file-system-over-etc/
https://community.toradex.com/t/automount-overlay-for-etc/15529/6
https://systemd-devel.freedesktop.narkive.com/XyE0Bz4o/dhcp6-client-failing-when-etc-is-mounted-as-overlayfs

The workaround for this issue (for now), is that someone has to manually
copy the value in /run/machine-id into /etc/machine-id on the
first boot. After that it works as normal. The workaround is basically to run
this command on the first boot:
$ cat /run/machine-id > /etc/machine-id

This will hopefully be fixed in a future release so that this workaround 
is no longer needed.