Add support for booting the ADSP to use the IMU (#12)
* Boot ADSP to enable IMU (#34)

This commit adds a systemd service and boot script to boot the ADSP
The SBC8550 has an IMU which is connected to the Snapdragon Sensor Core (SSC).
This service will boot the ADSP after /dev/remoteproc-adsp-md comes up.

* Bugfix for ADSP boot Service so it works as intended (#50)

This commit fixes a minor issue with the ADSP boot systemd service
where it's working, but not as intended. The service seems to start
regardless of the dependency specified in the service file
and this seems to be because the After= and Requires= clauses
are in the wrong section of the systemd service file.
After fixing this, I also noticed that systemd needs
us to create a device unit for /dev/remoteproc-adsp-md.
So I had a udev rule to create a sytemd device unit for
/dev/remoteproc-adsp-md.

This has been tested on the target and is working correctly.

* Added Sensor SSC ADSP loader driver to kernel configs

* Added DT node to enable SSC Sensors