Integrate Qualcomm AI Demos (#11)
Added a Yocto recipe which will build Qualcomm's AI demo's from
source. Each sample application has been split into a separate Yocto
package so that users can select which demos they want in their image.
And we have also created a packagegroup which includes
all the demos if they want to do that too. In our base image (imdt-image-weston)
we just use the packagegroup since our BSP will include all of demos.
Note that not all of the demos are working, so we are only building
and installing the demos we have tested that are working.

We have also kept the models out of the image because of potential
licensing issues. So users will need to install the demos manually
and in return accept the AI model license restrictions from Qualcomm.