Add support for TLV320AIC3101 for Speaker Playback with Pulseaudio and agmplay (#15)
- Add patches for kernel, audio-devicetree and audio-kernel. The
audio-devicetree requires the audio-kernel patches to be compiled
together because the DAI links have change in the header.

- Add new audio graph for the acdbdata to communicate with the DAC
through primary I2S. These files are binaries thus added here instead of
a separate repo.

- Patch agm-plugin to use 2 channels for the primary MI2S.

- Add audio service recipe to setup the correct controls for the DAC during boot and start pulseaudio.
The qtitinymix controls to setup the HP out were found after manually trying different chip configuration using I2C commands.
See: https://e2e.ti.com/support/audio-group/audio/f/audio-forum/1302645/tlv320aic3101-high-power-outputs-not-working
This link's register values was working so I verified with the driver what I need to do to enable this route.
The main issue was that DAC was not connected with the HP through the DAC_L2 route.
In general, TI's support forum has a lot of useful tickets for the chip.

- Modify pulseaudio configuration file to load the DAC as sink device and
connect the PAL_DEVICE_OUT_SPEAKER to the low-latency playback device.
The OUT_SPEAKER is connected with the DAC and the low-latency device is
the default one for playback. So when the pulseaudio starts, it will
load the DAC as the default playback device.

- Add mtp files for the qal recipe. qal is loading the mtp board as
default and the files were missing. The resourcemanager file is the same
with the others (ie resourcemanager_kalama_qrd.xml) but has the PRIMARY
MI2S lines as the PAL_DEVICE_OUT_SPEAKER, which enables the DAC. The
mixer file is the same with the other devices.

Changes are similar with another project,
apart from the acdbdata which needs other 
connections and the audio-init-service which has other controls.

The configuration has been tested with agmplay and pulseaudio
and it works when used with some headphones.