Fix eva kernel module symbol file mismatch issue on boot (#17)
The eva-kernel module recipe, builds a kernel
module which fails to load on the device.
When this module fails to load, it stops
other modules loading since the
systemd-module-load job/service
dies due to this failure. And it was failing to
load due to a symbol mismatch. This was simply
because they were not passing the arguments (KBUILD_EXTRA_SYMBOLS)
correctly to build_module.sh. With this fix,
the module now correctly loads with no errors.