Add support for updating the BSP image via SWUpdate (#18)
This commit adds support for updating the image via SWUpdate.
Our configuration of SWUpdate is a bit more barebones than
normal since Qualcomm/Android has their own tools for updating
(android recovery application) an image. So we are really just using
SWUpdate for hosting the webserver that users can upload the OTA image to
as a SWU file. And in our case, SWUpdate will also handle image verification.

We have also had to modify the swupdate systemd service
because the call to the android recovery application takes
a long time because our update images are large.
The SWUpdate systemd service had Type=notify which
meant that it sends status notifications from the SWUpdate
application to systemd. However, because the call
to the recovery application takes a while, it blocks
the execution and then causes the systemd service to
timeout and exit whilst updating the software.
So we have changed the service to Type=simple
to stop this from happening.

Our SWU image is just the update_ext4.zip which we pass
to the recovery application in the postinstall stage
of the update script (which is also contained in the SWU file).

We have also done some Yocto work here to separate the SWU image
build from the main image build. This is because Qualcomm
appends the tasks to generate the recovery image to the
main image (imdt-image-weston) build. However,
this can make all builds quite slow, as all image builds will
create the recovery image. We only need to create the recovery
image when building the SWU image. So we have added the
ENABLE_RECOVERY flag, which will selectively enable the
recovery image generation tasks. And we have done it in such
a way that this flag is set in the SWU image and thus the recovery
image generation only occurs when building the SWU image.