Enable SD Card using SDHCI 2 for V2 SBC HW (#24)
This commit adds the kernel and devicetree patches
to enable the SD card on V2 SBC HW. It will not work
properly on V1 HW as the 1.2V supply used for the
SD card level shifter will not be enabled. But it won't
break anything having this enabled on V1 HW. It will
detect the SD card insertion and will try to communicate
with it and fail.