Add minimal image for booting console and adb (#25)
This minimal image is great for bring-up as
it compiles no devicetree's or out-of-tree
drivers. So if we are working on just
device tree and kernel changes,
this is a good image to use for development.
It boots to console and also provides support
for ADB if the USB is available.