Add support for IW416 (onboard WiFi) connected to SDC4 (#28)
This commit adds support for the onboard WiFi chip (NXP IW416)
on the IMDT SBC8550. The main patches are:
- Patch to kalama devicetree to add SDHCI_4 node and cap the max speed to SDR25 as
SDR50 and above are not working.
- Patch to kernel that adds support for the IW416 chip in the mwifiex SDIO driver. This patch
has been backported from the upstream Linux mwifiex driver.

To actually use the chip we also need to load some firmware to it. This firmware
allows the IW416 to work as a WiFi and Bluetooth chip.
The firmware has been retrieved from this repo:
https://github.com/nxp-imx/imx-firmware/tree/lf-6.1.22_2.0.0